@@ -7,9 +7,9 @@
     "npm-run-all": "^4.1.2",
     "prettier": "^1.16.1",
     "prop-types": "^15.6.0",
-    "react": "^16.2.0",
+    "react": "^16.13.1",
     "react-bootstrap": "^0.32.1",
-    "react-dom": "^16.2.0",
+    "react-dom": "^16.13.1",
     "react-router-dom": "^4.2.2",
     "react-scripts": "1.1.1"
   },
